Move VICEAdminTabs into their own component and out of the page
